my_blog/requirements.txt
astroid==2.1.0
certifi==2018.11.29
colorama==0.4.1
confusable-homoglyphs==3.2.0
Django==2.1.5
django-comments-xtd==2.3.1
django-contrib-comments==1.9.0
django-crispy-forms==1.7.2
django-hitcount==1.3.0
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-registration==3.0
djangorestframework==3.9.1
docutils==0.14
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
markdown2==2.3.7
mccabe==0.6.1
Pillow==5.4.1
pylint==2.2.2
pytz==2018.7
six==1.12.0
wincertstore==0.2
wrapt==1.10.11
